* Fixed broken test. * Moved WaitForStateConsumption to internal package.
